Fix: Initial State of DataView
Bug: The DataView doesn't show the data
initially before running the algorithm.
This makes the impression that data is
not generated while we are changint the
data properties in the control panel.

Fix: use useEffects to look for changes
in dataLength, maxData and algorithmName
and whenever they change the algorithm
object gets recreated and the random
data is regenerated. There by showing
the random data with default values
before even tweaking the control panel
settings.